feat: split the page to small components.
